Code optimizations

+++ Easy/C/How Many Weeks.c
@@ -1,10 +1,8 @@
 #include <stdio.h>
 
-//Compiler version gcc  6.3.0
-
 int main() {
-	int n;
-	scanf("%d", &n);
-	printf("%d", 10000/n);
+	unsigned short n;
+	scanf("%hu", &n);
+	printf("%hu", 10000 / n);
 	return 0;
 }
